We only support Strings with less than INT_MAX characters

+++ src/main/c/cext/ruby.c
@@ -4421,7 +4421,7 @@ void rb_str_modify_expand(VALUE str, long expand) {
   if (expand < 0) {
     rb_raise(rb_eArgError, "negative expanding string size");
   }
-  if (expand > LONG_MAX - len) {
+  if (expand > INT_MAX - len) {
     rb_raise(rb_eArgError, "string size too big");
   }
